requirements.txt
streamlit
pandas
openpyxl
python-docx
beautifulsoup4
ipwhois
watchdog
